+++ lib/command/send.rb
@@ -41,6 +41,12 @@ module Command
       @opt.on("--without-freeze", "一括送信時に凍結された小説は対象外にする") {
         @options["without-freeze"] = true
       }
+      @opt.on("-b", "--backup-bookmark", "端末の栞データを手動でバックアップする(KindlePW系用)") {
+        @options["command-backup-bookmark"] = true
+      }
+      @opt.on("-r", "--restore-bookmark", "栞データのバックアップを端末にコピーする(KindlePW系用)") {
+        @options["command-restore-bookmark"] = true
+      }
     end
 
     def get_device(argv)
@@ -77,6 +83,16 @@ module Command
           titles[id] = data["title"]
         end
       end
+
+      case
+      when @options["command-backup-bookmark"]
+        process_backup_bookmark(device)
+        exit
+      when @options["command-restore-bookmark"]
+        process_restore_bookmark(device)
+        exit
+      end
+
       tagname_to_ids(argv)
       argv.each do |target|
         ebook_path = Narou.get_ebook_file_path(target, device.ebook_file_ext)
@@ -115,9 +131,36 @@ module Command
           exit Narou::EXIT_ERROR_CODE   # next しても次も失敗すると分かりきっているためここで終了する
         end
       end
+      if send_all && @options["backup-bookmark"]
+        process_backup_bookmark(device)
+      end
     rescue Device::SendFailure, Interrupt
       puts "送信を中断しました"
       exit Narou::EXIT_ERROR_CODE
+    end
+
+    def process_backup_bookmark(device)
+      device.extend(device.get_hook_module)
+      unless device.respond_to?(:backup_bookmark)
+        error "ご利用の端末での栞データのバックアップは対応していません"
+        return
+      end
+      if device.backup_bookmark > 0
+        puts "端末の栞データをバックアップしました"
+      end
+    end
+
+    def process_restore_bookmark(device)
+      device.extend(device.get_hook_module)
+      unless device.respond_to?(:restore_bookmark)
+        error "ご利用の端末での栞データのバックアップは対応していません"
+        return
+      end
+      if device.restore_bookmark > 0
+        puts "栞データを端末にコピーしました"
+      else
+        puts "栞データが無いようです"
+      end
     end
   end
 end